adjusting logic home and rename prototype component
